DataReceiver => ReadCallback, and improve semantics
