+++ tiledb/api/src/query/read/output.rs
@@ -135,23 +135,6 @@ pub trait HasScratchSpaceStrategy<C> {
     type Strategy: ScratchAllocator<C>;
 }
 
-pub trait DataReceiver: Sized {
-    type Unit: CAPISameRepr;
-
-    fn receive<'data>(
-        &mut self,
-        records: usize,
-        bytes: usize,
-        input: InputData<'data, Self::Unit>,
-    ) -> TileDBResult<()>;
-}
-
-pub trait ReadResult: Sized {
-    type Receiver: DataReceiver + Into<Self>;
-
-    fn new_receiver() -> Self::Receiver;
-}
-
 #[derive(Clone, Debug)]
 pub struct NonVarSized {
     pub capacity: usize,
@@ -243,42 +226,6 @@ where
     type Strategy = NonVarSized;
 }
 
-impl<C> DataReceiver for Vec<C>
-where
-    C: CAPISameRepr,
-{
-    type Unit = C;
-
-    fn receive<'data>(
-        &mut self,
-        records: usize,
-        _bytes: usize,
-        input: InputData<'data, C>,
-    ) -> TileDBResult<()> {
-        Ok(if let Buffer::Owned(data) = input.data {
-            if self.is_empty() {
-                *self = data.into_vec();
-                self.truncate(records)
-            } else {
-                self.extend_from_slice(&data[0..records])
-            }
-        } else {
-            self.extend_from_slice(&input.data.as_ref()[0..records])
-        })
-    }
-}
-
-impl<C> ReadResult for Vec<C>
-where
-    C: CAPISameRepr,
-{
-    type Receiver = Self;
-
-    fn new_receiver() -> Self::Receiver {
-        vec![]
-    }
-}
-
 impl HasScratchSpaceStrategy<u8> for Vec<String> {
     type Strategy = VarSized;
 }
@@ -358,220 +305,6 @@ impl<'data, C> Iterator for VarDataIterator<'data, C> {
             Some(&data_buffer[start..start + slen])
         } else {
             None
-        }
-    }
-}
-
-impl DataReceiver for Vec<String> {
-    type Unit = u8;
-
-    fn receive<'data>(
-        &mut self,
-        records: usize,
-        bytes: usize,
-        input: InputData<'data, Self::Unit>,
-    ) -> TileDBResult<()> {
-        for s in VarDataIterator::new(records, bytes, &input)? {
-            let s = String::from_utf8_lossy(s);
-            self.push(s.to_string());
-        }
-
-        Ok(())
-    }
-}
-
-impl ReadResult for Vec<String> {
-    type Receiver = Self;
-
-    fn new_receiver() -> Self::Receiver {
-        vec![]
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use proptest::collection::vec;
-    use proptest::prelude::*;
-
-    const MIN_RECORDS: usize = 0;
-    const MAX_RECORDS: usize = 1024;
-
-    const MIN_BYTE_CAPACITY: usize = 0;
-    const MAX_BYTE_CAPACITY: usize = 1024 * 1024;
-
-    fn do_read_result_repr<C>(dst_unit_capacity: usize, unitsrc: Vec<C>)
-    where
-        C: CAPISameRepr + std::fmt::Debug + PartialEq,
-    {
-        let alloc = <NonVarSized as ScratchAllocator<C>>::construct(Some(
-            dst_unit_capacity,
-        ));
-
-        let mut scratch_space = alloc.scratch_space();
-
-        let mut unitdst = <Vec<C> as ReadResult>::new_receiver();
-
-        while unitdst.len() < unitsrc.len() {
-            let ncells = std::cmp::min(
-                scratch_space.0.len(),
-                unitsrc.len() - unitdst.len(),
-            );
-            if ncells == 0 {
-                scratch_space = alloc.realloc(scratch_space);
-                continue;
-            }
-
-            unsafe {
-                std::ptr::copy_nonoverlapping::<C>(
-                    unitsrc[unitdst.len()..unitsrc.len()].as_ptr(),
-                    scratch_space.0.as_mut_ptr(),
-                    ncells,
-                )
-            };
-
-            let input_data = InputData {
-                data: Buffer::Borrowed(&*scratch_space.0),
-                cell_offsets: None,
-            };
-
-            let prev_len = unitdst.len();
-
-            <Vec<C> as DataReceiver>::receive(
-                &mut unitdst,
-                ncells,
-                ncells * std::mem::size_of::<u64>(),
-                input_data,
-            )
-            .expect("Error aggregating data into Vec");
-
-            assert_eq!(ncells, unitdst.len() - prev_len);
-            assert_eq!(unitsrc[0..unitdst.len()], unitdst);
-        }
-
-        assert_eq!(unitsrc, unitdst);
-    }
-
-    proptest! {
-        #[test]
-        fn read_result_u64(dst_unit_capacity in MIN_RECORDS..=MAX_RECORDS, unitsrc in vec(any::<u64>(), MIN_RECORDS..=MAX_RECORDS)) {
-            do_read_result_repr::<u64>(dst_unit_capacity, unitsrc)
-        }
-
-        #[test]
-        fn read_result_u32(dst_unit_capacity in MIN_RECORDS..=MAX_RECORDS, unitsrc in vec(any::<u32>(), MIN_RECORDS..=MAX_RECORDS)) {
-            do_read_result_repr::<u32>(dst_unit_capacity, unitsrc)
-        }
-
-        #[test]
-        fn read_result_u16(dst_unit_capacity in MIN_RECORDS..=MAX_RECORDS, unitsrc in vec(any::<u16>(), MIN_RECORDS..=MAX_RECORDS)) {
-            do_read_result_repr::<u16>(dst_unit_capacity, unitsrc)
-        }
-
-        #[test]
-        fn read_result_u8(dst_unit_capacity in MIN_RECORDS..=MAX_RECORDS, unitsrc in vec(any::<u8>(), MIN_RECORDS..=MAX_RECORDS)) {
-            do_read_result_repr::<u8>(dst_unit_capacity, unitsrc)
-        }
-
-        #[test]
-        fn read_result_f64(dst_unit_capacity in MIN_RECORDS..=MAX_RECORDS, unitsrc in vec(any::<f64>(), MIN_RECORDS..=MAX_RECORDS)) {
-            do_read_result_repr::<f64>(dst_unit_capacity, unitsrc)
-        }
-
-        #[test]
-        fn read_result_f32(dst_unit_capacity in MIN_RECORDS..=MAX_RECORDS, unitsrc in vec(any::<f32>(), MIN_RECORDS..=MAX_RECORDS)) {
-            do_read_result_repr::<f32>(dst_unit_capacity, unitsrc)
-        }
-    }
-
-    fn do_read_result_strings(
-        record_capacity: usize,
-        byte_capacity: usize,
-        stringsrc: Vec<String>,
-    ) {
-        let alloc = VarSized {
-            byte_capacity,
-            offset_capacity: record_capacity,
-        };
-
-        let mut scratch_space = alloc.scratch_space();
-
-        let mut stringdst: Vec<String> = vec![];
-
-        while stringdst.len() < stringsrc.len() {
-            /* copy from stringsrc to scratch data */
-            let (nrecords, nbytes) = {
-                /* write the offsets first */
-                let (nrecords, nbytes) = {
-                    let mut scratch_offsets = scratch_space.1.as_mut().unwrap();
-                    let mut i = 0;
-                    let mut off = 0;
-                    let mut src =
-                        stringsrc[stringdst.len()..stringsrc.len()].iter();
-                    loop {
-                        if i >= scratch_offsets.len() {
-                            break (i, off);
-                        }
-                        if let Some(src) = src.next() {
-                            if off + src.len() <= scratch_space.0.len() {
-                                scratch_offsets[i] = off as u64;
-                                off += src.len();
-                            } else {
-                                break (i, off);
-                            }
-                        } else {
-                            break (i, off);
-                        }
-                        i += 1;
-                    }
-                };
-
-                if nrecords == 0 {
-                    assert_eq!(0, nbytes);
-                    scratch_space = alloc.realloc(scratch_space);
-                    continue;
-                }
-
-                let scratch_offsets = scratch_space.1.as_ref().unwrap();
-
-                /* then transfer contents */
-                for i in 0..nrecords {
-                    let s = &stringsrc[stringdst.len() + i];
-                    let start = scratch_offsets[i] as usize;
-                    let end = if i + 1 < nrecords {
-                        scratch_offsets[i + 1] as usize
-                    } else {
-                        nbytes
-                    };
-                    scratch_space.0[start..end].copy_from_slice(s.as_bytes())
-                }
-
-                (nrecords, nbytes)
-            };
-
-            /* then copy from scratch data to stringdst */
-            let prev_len = stringdst.len();
-            let input = InputData {
-                data: Buffer::Borrowed(&scratch_space.0),
-                cell_offsets: scratch_space
-                    .1
-                    .as_ref()
-                    .map(|c| Buffer::Borrowed(c)),
-            };
-            stringdst
-                .receive(nrecords, nbytes, input)
-                .expect("Error aggregating Vec<String>");
-
-            assert_eq!(nrecords, stringdst.len() - prev_len);
-            assert_eq!(stringsrc[0..stringdst.len()], stringdst);
-        }
-    }
-
-    proptest! {
-        #[test]
-        fn read_result_strings(record_capacity in MIN_RECORDS..=MAX_RECORDS, byte_capacity in MIN_BYTE_CAPACITY..=MAX_BYTE_CAPACITY, stringsrc in vec(any::<String>(), MIN_RECORDS..=MAX_RECORDS))
-        {
-            do_read_result_strings(record_capacity, byte_capacity, stringsrc)
         }
     }
 }
